Move generalized components
