+++ src/containers/SearchableList/SearchableList.js
@@ -6,9 +6,9 @@ import { Button } from '@codaco/ui';
 import SearchIcon from '@material-ui/icons/SearchRounded';
 import Loading from '../../components/Loading';
 import Panel from '../../components/Panel';
+import useSort from '../../hooks/useSort';
+import useSearch from '../../hooks/useSearch';
 import useSortableProperties from './useSortableProperties';
-import useSort from './useSort';
-import useSearch from './useSearch';
 import HyperList from '../HyperList';
 
 const modes = {
