+++ src/hooks/useSearch.js
@@ -5,7 +5,7 @@ import {
   useRef,
 } from 'react';
 import Fuse from 'fuse.js';
-import useDebounce from '../../hooks/useDebounce';
+import useDebounce from './useDebounce';
 
 const MIN_QUERY_LENGTH = 2;
 const SEARCH_DELAY = 1000;
